Add guard against unrecognized arg
Previously, cli arguments were being mapped into an unordered_map, which
had the side effect of storing args out of order. This prevented using
guards in if/else chains in some cases, the most important of which was
when checking for unrecognized argumens. Unordered_map was replaced with
map which fixes this issue.